Move Grid input to corecel/inp
