+++ app/celer-g4/RunInputIO.json.cc
@@ -7,6 +7,7 @@
 #include "RunInputIO.json.hh"
 
 #include "corecel/cont/ArrayIO.json.hh"
+#include "corecel/grid/GridTypesIO.json.hh"
 #include "corecel/io/JsonUtils.json.hh"
 #include "corecel/io/Logger.hh"
 #include "corecel/io/StringEnumMapper.hh"
