+++ src/celeritas/Types.cc
@@ -104,18 +104,4 @@ char const* to_cstring(WlsTimeProfile value)
 }
 
 //---------------------------------------------------------------------------//
-/*!
- * Get a string corresponding to the interpolation method.
- */
-char const* to_cstring(InterpolationType value)
-{
-    static EnumStringMapper<InterpolationType> const to_cstring_impl{
-        "linear",
-        "poly_spline",
-        "cubic_spline",
-    };
-    return to_cstring_impl(value);
-}
-
-//---------------------------------------------------------------------------//
 }  // namespace celeritas
