+++ src/celeritas/TypesIO.json.hh
@@ -16,8 +16,5 @@ namespace celeritas
 void from_json(nlohmann::json const& j, TrackOrder& value);
 void to_json(nlohmann::json& j, TrackOrder const& value);
 
-void from_json(nlohmann::json const& j, InterpolationType& value);
-void to_json(nlohmann::json& j, InterpolationType const& value);
-
 //---------------------------------------------------------------------------//
 }  // namespace celeritas
