+++ src/celeritas/io/ImportOpticalModel.hh
@@ -8,7 +8,7 @@
 
 #include <vector>
 
-#include "celeritas/inp/Grid.hh"
+#include "corecel/inp/Grid.hh"
 
 namespace celeritas
 {
