+++ src/celeritas/optical/model/MieModel.cc
@@ -11,9 +11,9 @@
 #include "corecel/Assert.hh"
 #include "corecel/Types.hh"
 #include "corecel/data/CollectionBuilder.hh"
+#include "corecel/inp/Grid.hh"
 #include "corecel/io/Logger.hh"
 #include "celeritas/Types.hh"
-#include "celeritas/inp/Grid.hh"
 #include "celeritas/optical/CoreParams.hh"
 #include "celeritas/optical/CoreState.hh"
 #include "celeritas/optical/InteractionApplier.hh"
