+++ src/celeritas/phys/detail/EnergyMaxXsCalculator.hh
@@ -6,7 +6,7 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
-#include "celeritas/inp/Grid.hh"
+#include "corecel/inp/Grid.hh"
 #include "celeritas/phys/PhysicsOptions.hh"
 #include "celeritas/phys/Process.hh"
