+++ src/corecel/grid/GridTypesIO.json.hh
@@ -2,30 +2,19 @@
 // Copyright Celeritas contributors: see top-level COPYRIGHT file for details
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/io/ImportPhysicsTable.hh
+//! \file corecel/grid/GridTypesIO.json.hh
 //---------------------------------------------------------------------------//
 #pragma once
 
-#include <vector>
+#include <nlohmann/json.hpp>
 
-#include "corecel/inp/Grid.hh"
-
-#include "ImportUnits.hh"
+#include "GridTypes.hh"
 
 namespace celeritas
 {
 //---------------------------------------------------------------------------//
-/*!
- * Imported physics table. Each table stores physics vectors for all materials.
- */
-struct ImportPhysicsTable
-{
-    ImportUnits x_units{ImportUnits::unitless};
-    ImportUnits y_units{ImportUnits::unitless};
-    std::vector<inp::UniformGrid> grids;
-
-    explicit operator bool() const { return !grids.empty(); }
-};
+void from_json(nlohmann::json const& j, InterpolationType& value);
+void to_json(nlohmann::json& j, InterpolationType const& value);
 
 //---------------------------------------------------------------------------//
 }  // namespace celeritas
